Refactored code;
